Change hostNetworkConfiguration API
To ensure that traffic from pod to external works when routed HBN is
used, we need to know upfront the mapping between the name of the node
on the host cluster and the name of the node in the DPU cluster. To make
it easier for the user to create the configuration for each host one by
one and to simplify the discovery in code, I've changed the API to make
the configuration more specific to each host.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>