Ensure ArgoCD, kamaji-etcd and NFD pods can start when VF injector is installed
Given that we do soft affinity on the control plane nodes and not hard,
the VF injector injects a VF. This is problematic tho because at the
deployment phase, we only have control plane nodes. To avoid this
situation, we add a label to the pods to ignore the VF injection. This
has the risk that if the pods ever get scheduled on the workers, they
won't be able to run.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>