feat: add dpucluster, dpuset and dpu to dpfctl describe
* fix: remove pointer from ObjectTree in scope

We don't need to have the ObjectTree a pointer in our scope

Signed-off-by: Tobias Giese <tgiese@nvidia.com>

* fix: remove conditions from dpuset and conditions methods

We do not longer need the DPUSet conditions. Additionally, the methods GetConditions
and SetConditions are also no longer necessary for the provisioning CRDs.

Signed-off-by: Tobias Giese <tgiese@nvidia.com>

* feat: add dpucluster, dpuset and dpu to dpfctl describe

This commit adds the DPUCluster, DPUSets and DPUs to dpfctl describe.
Additionally, to get dpfctl work with DPUSets we had to add conditions to the DPUSet CRD.

Signed-off-by: Tobias Giese <tgiese@nvidia.com>