fix: add no namespace to conditions in dpfctl
* fix: add no namespace to conditions in dpfctl

There was an issue with the status conditions in dpfctl. As the new namespace column
has been introduced the namespace was not added to the conditions rows.
To fix this issue we should add an empty string to the namespace column.
Additionally, the addResourceByGVK() will be skipped if the resource shouldn't be displayed.

Signed-off-by: Tobias Giese <tgiese@nvidia.com>