chore: set BackoffLimit explicitly for job dowloader
* chore: set BackoffLimit explicitly for job dowloader

Fail a Job after some amount of retries due to a error in configuration, pod etc.
To do so, set .spec.backoffLimit to specify the number of retries before considering
a Job as failed.
The back-off limit is set by default to 6.

This field cannot be used in combination  with restartPolicy=OnFailure.

Signed-off-by: Igor Ivanov <igori@nvidia.com>