feat: add filter to show only failed conditions to dpfctl describe
* feat: add filter to show only failed conditions to dpfctl describe

With the show-conditions flag it is possible to filter certain kinds or kind/name.
But in many cases it is helpful to show only failed conditions.

To enable this feature we can simply check in the isObjDebug() function if the filter is set to "failed"
and the objects Ready condition is False.

As we are using the isObjDebug() function for the expand-resources flag as well, we also enabled this
new feature for the expand-resources flag.

Signed-off-by: Tobias Giese <tgiese@nvidia.com>