fix: ensure flannel doesn't write CNI conf file
* fix: ensure flannel doesn't write CNI conf file

There is a race condition with multus and flannel. When a new
DPU node joins the cluster, there may be a case where flannel
writes the CNI conf in /etc/cni/net.d before multus does. This
can happen if the flannel pod started before the multus one.
This means that Pods with multus annotations (e.g. dummy DPUService
pods) may bypass multus (since multus didn't write its own CNI
conf yet) and use flannel directly instead. This means that
the secondary network will not be injected at all. On delete
of the pod, multus already has written the CNI conf which means
that the cniDel will go through multus. Multus will delegate the
cniDel to NVIPAM since the NAD used specifies nvipam as ipam plugin,
however, the annotation won't have the poolName injected by the Pod
IPAM Injector since the SFC controller binary that includes the aforementioned
controller may have lost the window to inject the correct pool name
because the dummy DPUService pod didn't stay in Pending since flannel
was the primary CNI at that time (remember that multus didn't manage
to write its own cni conf yet). The CNI call will fail with a message
related to pool not found.

With this patch, we ensure that flannel won't write any CNI conf and we
wait for multus to to write it for the first time.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>