fix: use object name as node label in dpudeployment upgrades
* fix: use object name as node label

This PR switch from digest to object name for placing objects, i.e dpuService, dpuServiceInterface or dpuServiceChain on dpu nodes.
We expect an object name to be unique in the namespace at a given, this gives us the ability to label the node for only one object at a given time.

Signed-off-by: Soule BA <souleb@nvidia.com>