fix: assign nodeAffinity on the CP for the servicechainset
* fix: assign nodeAffinity on the CP for the servicechainset

When deploying OVN Kubernetes, we install the VF injector as well. If we
don't set affinities, the servicechainset controller will get a VF and
will not be pinned on the control plane. This commit fixes that.

Usually for setting tolerations and affinities we use the edits
functionality of the internal/operator/inventory to make them kinda
testable. However, here I see that we used tolerations directly in the
chart that's why I went with the same approach for affinities.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>