fix: always patch the current version of child resources
* chore: rename function

rename reconcileCurrentRevision to reconcileCurrentDPUServiceChainRevision

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>

* fix: always set managedFields and GVK before patching objects

always set GVK and remove managedFields to ensure the patch can go
through just before patching. This commit fixes a bug when unpausing
DPUServices where the patch would fail because the managedFields are
set.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>

* fix: always patch the current version of child resources

Previously, we didn't patch the current versions of DPUServiceChain,
DPUServiceInterface and DPUService resources. This can lead to several
issues like a user modifying the resource manually would leave the
resource in a bad state and also a race condition within the controller
itself that has to do with eventual consistency between mutating API
calls to Kubernetes API server and the informer updating the internal
controller runtime caches.

To avoid all of the above, we now patch the most current resources. The
user can still break the setup by modifying the annotation that contains
the version in each of these resources but the system should be able to
recover by creating a new application and eventually getting rid of the
old one.

Signed-off-by: Vasilis Remmas <vremmas@nvidia.com>