fix: exposeDevice and destroyDevice functions
* test: storage_vendor_dpu_rpc and snap_rpc tests

Unit tests for the functions in storage_vendor_dpu_rpc
and snap_rpc, ensuring proper functionality.


* fix: minimize rpc calls

Instead of calling emulation_function_list, nvme_subsystem_list,
and bdev_get_bdevs, bdev_nvme_get_controllers RPCs in each function
separately, call them once at the beginning and pass the retrieved
lists to the functions that need them.


* fix: handleDetachment function

We encountered an issue in handleDetachment()
where the function returned before reaching removeFinalizer().
As a result, the volumeAttachment could not be deleted,
causing handleDetachment() to be called repeatedly in a loop.
To resolve this, we moved removeFinalizer() to execute after
handleDetachment() returns, ensuring that the finalizer is
removed and the volumeAttachment is properly deleted.


* fix: exposeDevice and destroyDevice functions

Added checks to the ExposeDevice and DestroyDevice
functions to ensure they are not called multiple
times by the controller reconcile loop,
preventing redundant operations and failures.
Change BdevNvmeDetachControllerRequest to get
controller name.

fix: #4245779 & #4245773

Signed-off-by: Linoy Halfon <lhalfon@nvidia.com>